+++ durabletask/client.py
@@ -104,7 +104,8 @@ class TaskHubGrpcClient:
     def schedule_new_orchestration(self, orchestrator: Union[task.Orchestrator[TInput, TOutput], str], *,
                                    input: Union[TInput, None] = None,
                                    instance_id: Union[str, None] = None,
-                                   start_at: Union[datetime, None] = None) -> str:
+                                   start_at: Union[datetime, None] = None,
+                                   reuse_id_policy: Union[pb.OrchestrationIdReusePolicy, None] = None) -> str:
 
         name = orchestrator if isinstance(orchestrator, str) else task.get_name(orchestrator)
 
@@ -113,7 +114,9 @@ class TaskHubGrpcClient:
             instanceId=instance_id if instance_id else uuid.uuid4().hex,
             input=wrappers_pb2.StringValue(value=shared.to_json(input)) if input is not None else None,
             scheduledStartTimestamp=helpers.new_timestamp(start_at) if start_at else None,
-            version=wrappers_pb2.StringValue(value=""))
+            version=wrappers_pb2.StringValue(value=""),
+            orchestrationIdReusePolicy=reuse_id_policy,
+            )
 
         self._logger.info(f"Starting new '{name}' instance with ID = '{req.instanceId}'.")
         res: pb.CreateInstanceResponse = self._stub.StartInstance(req)